Agent/requirements.txt
google-cloud-aiplatform
fastapi
uvicorn
pydantic
pydantic-settings
python-dotenv
rich
